move profiling scripts to scripts/profiling + add doc
